move the helper code into testing_utils
